feat(luna-test,devtools,docs,lib,tools): luna v1.22.0 pinned; fixed-point asserts, cppcheck in lint, generated luna reference, in-flight branches cleared
Four Tier 2 items of the gaps review plus the luna pin bump, one commit
because they share the validation run.

luna v1.22.0 (R-B of the 2026-09-11 report): `luna profile` rows carry
`per_frame {max, max_frame, mean, frames}` and `--budget SYMBOL=MCLK`
gates on the per-frame max. Nothing in emulation moved: the suite is
green on the pin with no re-baseline. The VBlank gate (R4) waits on one
step — rows are per label and the NMI handler is five of them, so a
budget on `NmiHandler` judges its entry block only; luna is asked to
fold `@child` rows into their parent (~/opensnes_reports/, owner-side).

L2b — `devtools/libtests` gains 19 runtime vectors for fixMul, fixDiv,
fixLerp (8.8) and fix32Mul, fix32Div (16.16): signs, fractions, negative
deltas, t = 0 / 128 / 255, fixDiv's zero-divisor contract, 1/3 truncated.
The first build found `FIX(x)` undefined for a negative argument — the
public macro shifted `(x) << 8`, which the clang pre-pass rejects, while
docs/tutorials/math.md itself writes `FIX(-3)`. The shift is unsigned
now; same bits, corpus byte-identical.

H4 — `make lint-cppcheck`, in `make lint`, cppcheck installed in the CI
lint job: the asset tools' sources and the lib's C gate, lodepng and
stb_image suppressed, the w65816 backend advisory. First run, tools:
cmdparser (both vendored copies) kept `ctx` pointing at a block-local
context after the block; aseprite2snes read `ticks[0]` uninitialised for
a tag with no frames; smconv and img2snes grew buffers with
`p = realloc(p, …)`, which leaks the buffer on failure (now a checked
temporary that exits on NULL); smconv's io_write_asciif tested `str[i]`
before `i < length`; ten `%i`/`%d` printed unsigned sizes; gfx4snes's
free-then-fatal error paths and palplan's die() paths read as double
frees until both were declared noreturn (they gain printf format
checking too). One suppression with its reason: gfx4snes's palette merge
encodes a final's index as `num - 100` and recovers it, so the flagged
`final[n]` is in range. lib: the audio blob's two labels are one
section, suppressed inline. Four `%d` of unsigned ids in the emitter's
debug comments fixed; its other six advisory hits are upstream idioms.

D3 — docs/tools/luna.md is generated from the pinned binary's own
--help (devtools/gen_luna_doc.py, 11 subcommands, stamped with the pin);
`make tests` fails when the page and the installed pin disagree, the
lint job (no luna) skips it. Linked from the toolbox index.

Pin: compiler/qbe → 1422c17 (62 patches, the emitter comment fix).

D7 — the five `wip/*` chantiers the review named no longer exist; the
four `wip/*` on origin were superseded by develop (three CI hygiene
branches, one harness branch) and are deleted. ROADMAP gains a "Work in
flight" section that says the plan is this backlog.

+++ lib/include/snes/math.h
@@ -66,9 +66,14 @@ typedef s16 fixed;
  * @code
  * fixed pos = FIX(50);   // 50.0 in fixed-point
  * fixed half = FIX(1) / 2;  // 0.5
+ * fixed back = FIX(-3);  // -3.0
  * @endcode
+ *
+ * @note The shift is done unsigned: `(-3) << 8` is undefined in C and the
+ *   build's clang pre-pass rejects it, so the value goes through u16 and
+ *   back. Same bits, same code, no diagnostic for a negative argument.
  */
-#define FIX(x) ((fixed)((x) << 8))
+#define FIX(x) ((fixed)(s16)((u16)(x) << 8))
 
 /**
  * @brief Convert fixed-point to integer (truncate)
